# openapi.yaml
openapi: 3.0.1
info:
  title: S3 Access
  description: GPT-4 browser Plugin to access, read, write, delete and modify files in your S3 bucket.
  version: 'v1'
servers:
  - url: http://localhost:5003
paths:
  /bucket/{bucketName}:
    get:
      operationId: getBucket
      summary: Read S3 Bucket
      parameters:
        - name: bucketName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid bucket name supplied
        '404':
          description: Bucket not found
  /bucket/{bucketName}/{folderName}:
    get:
      operationId: getBucketFolder
      summary: Read S3 Bucket Folder
      parameters:
        - name: bucketName
          in: path
          required: true
          schema:
            type: string
        - name: folderName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid bucket or folder name supplied
        '404':
          description: Bucket or folder not found
  /bucket/{bucketName}/{folderName}/{fileName}:
    get:
      operationId: getBucketFile
      summary: Read S3 Bucket File
      parameters:
        - name: bucketName
          in: path
          required: true
          schema:
            type: string
        - name: folderName
          in: path
          required: true
          schema:
            type: string
        - name: fileName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid bucket, folder or file name supplied
        '404':
          description: Bucket, folder or file not found
    post:
      operationId: createBucketFile
      summary: Write S3 Bucket File
      parameters:
        - name: bucketName
          in: path
          required: true
          schema:
            type: string
        - name: folderName
          in: path
          required: true
          schema:
            type: string
        - name: fileName
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true  # This allows any type of JSON object
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid bucket, folder or file name supplied
        '404':
          description: Bucket or folder not found
    put:
      operationId: updateBucketFile
      summary: Modify S3 Bucket File
      parameters:
        - name: bucketName
          in: path
          required: true
          schema:
            type: string
        - name: folderName
          in: path
          required: true
          schema:
            type: string
        - name: fileName
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true  # This allows any type of JSON object
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid bucket, folder or file name supplied
        '404':
          description: Bucket, folder or file not found
    delete:
      operationId: deleteBucketFile
      summary: Delete S3 Bucket File
      parameters:
        - name: bucketName
          in: path
          required: true
          schema:
            type: string
        - name: folderName
          in: path
          required: true
          schema:
            type: string
        - name: fileName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid bucket, folder or file name supplied
        '404':
          description: Bucket, folder or file not found